requirements.txt
fastapi
uvicorn[standard]
python-dotenv
SQLAlchemy
mysqlclient
aiomysql
PyMySQL
aiokafka
motor
beanie
pydantic-settings
cryptography
aiofiles
python-multipart